bacnet-stack: Update to revision 0.8.2; adjust Makefiles
Update to the latest released version 0.8.2 of the bacnet-stack.
This recipe grabs a specific revision from the 0.8.0 branch that
fixes a build issue with the gateway demo.

Re-adjust the Makefiles to handle changes that have occurred since
the 0.7.1 release.  The static library, demos and gateway were
confirmed to build under a Linux cross-compile environment.

Signed-off-by: Bryan Evenson <bevenson@melinkcorp.com>